Change VeraidAuthClient.run() to VeraidAuthClientMaker.get(), and convert UserCreationCommand to a function
